Refactor: useReducer / store
useReducer 사용하여 리팩토링함.
store를 분리하여 코드를 간결하게 변경
